# Conflicts:
#	core/src/main/java/com/IONA/TowerDefense/model/Enemy.java
#	core/src/main/java/com/IONA/TowerDefense/model/Path.java
#	core/src/main/java/com/IONA/TowerDefense/model/WayPoint.java

+++ core/src/main/java/com/IONA/TowerDefense/model/Movable.java
@@ -6,7 +6,7 @@ public interface Movable {
 
     Point getPosition();
 
-    float getSpeed();
+    int getSpeed();
 
     void setPosition(Point position);
